backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMuPDF==1.27.2.3
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
python-pptx==0.6.21
Pillow==11.0.0
reportlab==4.2.5
python-dotenv==1.0.0
gunicorn==21.2.0
